requirement.txt
appdirs==1.4.4
audeer==1.18.0
audiofile==1.1.0
audioread==3.0.0
audtorch==0.6.4
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.11.0
cytoolz==0.11.0
dask==2021.3.0
dataclasses==0.8
decorator==5.1.1
editdistance==0.5.3
fastdtw==0.3.4
idna==3.4
imageio==2.9.0
importlib-resources==5.4.0
joblib==1.1.1
kiwisolver==1.3.1
librosa==0.9.2
llvmlite==0.36.0
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.5
numba==0.53.1
numpy==1.19.5
olefile==0.46
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
pooch==1.6.0
pycparser==2.21
pyparsing==3.0.4
pyro-api==0.1.2
pyro-ppl==1.6.0
Pyro4==4.82
python-dateutil==2.8.2
pytz==2022.4
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.27.1
resampy==0.4.2
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.2
serpent==1.41
six==1.16.0
soundfile==0.11.0
sox==1.4.1
tabulate==0.8.10
threadpoolctl==3.1.0
tifffile==2020.10.1
toolz==0.11.2
torch==1.8.1+cu101
torchaudio==0.8.1
torchvision==0.9.1+cu101
tornado==6.1
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.12
xlrd==2.0.1
zipp==3.6.0
